requirements.txt
streamlit
extra_streamlit_components
streamlit_option_menu
Pillow
streamlit-extras
